% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringCoefficient.R
\name{clusteringCoefficient}
\alias{clusteringCoefficient}
\title{Calculate the clustering coefficient}
\usage{
clusteringCoefficient(adj, FUN = "localClusteringCoefficient", ...)
}
\arguments{
\item{adj}{An adjacency matrix. Calculating the clustering coefficient only makes sense if some connections are zero i.e. no connection.}

\item{FUN}{The function for calculating the clustering coefficient.}

\item{...}{Arguments to pass to FUN}
}
\value{
The clustering coefficient(s) for the adjacency matrix.
}
\description{
Calculate the clustering coefficient for an adjacency matrix. By default, the local clustering coefficient is calculated. From the PCIT package after it was archived on the CRAN.
}
\examples{
clusteringCoefficient(network_gen(50,.33)$network)
}
\seealso{
\link{localClusteringCoefficient}
}
\author{
Nathan S. Watson-Haigh
}
