% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{netsimul}
\alias{netsimul}
\title{Simulated network}
\format{
A list of three objects :
\describe{
  \item{number_genes}{The number of genes in the network}
  \item{clust_coef}{The clustering coefficient}
  \item{network}{The simulated network}
  }
}
\usage{
netsimul
}
\description{
Result of the use of the \code{network_gen} function.
}
\keyword{datasets}
