% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_gen.R
\name{network_gen}
\alias{network_gen}
\title{Random scale-free network generation. This function is used intensively in the abc function.}
\usage{
network_gen(number_genes, clust_coef)
}
\arguments{
\item{number_genes}{A number}

\item{clust_coef}{A number}
}
\value{
A list with the number of of genes, the targeted clustering coefficient and the resulting network
}
\description{
Generate random network topology
}
\examples{
network_gen(10,1)
}
