% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randABC-package.R
\docType{package}
\name{networkABC}
\alias{networkABC}
\title{networkABC}
\description{
An inference tool based on approximate Bayesian computations to decipher 
network data and assess the strength of their inferred links.
}
\references{
\emph{networkABC: An inference tool for networks based on approximate Bayesian computation}, Myriam Maumy-Bertrand, Frédéric Bertrand, preprint.
}
