% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{resabc}
\alias{resabc}
\title{Result of an ABC inference}
\format{
A list of 14 objects :
\describe{
\item{data}{: The microarray data used, rows are genes and columns are time points.)}
\item{ngenes}{: The number of genes.)}
\item{ntimes}{: The number of timepoints)}
\item{clust_size}{: the size of clusters}
\item{clust_coeffs}{: the clustering coefficient}
\item{tolerance}{: the tolerance between the generated networks and the reference network}
\item{number_hubs}{: number of hubs in the network}
\item{iterations}{: number of times to repeat ABC algorithm}
\item{number_networks}{: number of generated networks in each iteration of the ABC algorithm}
\item{hub_probs}{: one-dimensional array of size number_genes for the each label to be in the role of a hub}
\item{neighbour_probs}{: matrix of neighbour probabilities of size number_nodes*number_nodes}
\item{is_probs}{: is equal to 1 since hub_probs and neighbour_probs were specified}
}
}
\usage{
resabc
}
\description{
Result for the reverse engineering of a simulated Cascade network
}
\keyword{datasets}
