% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showHp}
\alias{showHp}
\title{Plot for the hub probabilities}
\usage{
showHp(result)
}
\arguments{
\item{result}{: The result of the abc algorithm.}
}
\description{
Plot for the hub probabilities ; there is one probability for each node in the network.
}
\examples{
data(resabc)
showHp(resabc)
}
