% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showNetwork}
\alias{showNetwork}
\title{Plot the final network.}
\usage{
showNetwork(res, min_prob)
}
\arguments{
\item{res}{: The result of the abc algorithm.}

\item{min_prob}{: numeric ; under this probabilitie value, the link between two genes is set to 0.}
}
\description{
Plot the final network.
}
\examples{
data(resabc)
showNetwork(resabc,.2)
}
