% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showNp}
\alias{showNp}
\title{Plot for the neighbourhood probabilities}
\usage{
showNp(result)
}
\arguments{
\item{result}{: The result of the abc algorithm.}
}
\description{
Plot for the neighbourhood probabilities ; there is one probability for each pair of node in the network.
}
\examples{
data(resabc)
showNp(resabc)
}
